+++ engine/engine/enclave_manager/enclave_manager.go
@@ -16,6 +16,8 @@ import (
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"net"
+	"strings"
+	"sync"
 	"time"
 )
 
@@ -36,10 +38,17 @@ const (
 	availabilityWaiterBinaryFilepath = "/run/api-container-availability-waiter"
 
 	shouldFetchStoppedContainers = false
+
+	// We set this to true in case there are any race conditions with a container starting as we're trying to stop the enclave
+	shouldKillAlreadyStoppedContainersWhenStoppingEnclave = true
 )
 
 // Manages Kurtosis enclaves, and creates new ones in response to running tasks
 type EnclaveManager struct {
+	// We use Docker as our backing datastore, but it has tons of race conditions so we use this mutex to ensure
+	//  enclave modifications are atomic
+	mutex *sync.Mutex
+	
 	dockerManager *docker_manager.DockerManager
 
 	dockerNetworkAllocator *docker_network_allocator.DockerNetworkAllocator
@@ -48,7 +57,8 @@ type EnclaveManager struct {
 func NewEnclaveManager(dockerManager *docker_manager.DockerManager) *EnclaveManager {
 	dockerNetworkAllocator := docker_network_allocator.NewDockerNetworkAllocator(dockerManager)
 	return &EnclaveManager{
-		dockerManager:           dockerManager,
+		mutex:                  &sync.Mutex{},
+		dockerManager:          dockerManager,
 		dockerNetworkAllocator: dockerNetworkAllocator,
 	}
 }
@@ -62,6 +72,9 @@ func (manager *EnclaveManager) CreateEnclave(
 	enclaveId string,
 	isPartitioningEnabled bool,
 	shouldPublishAllPorts bool) (string, *net.IPNet, string, *net.IP, *nat.PortBinding, error) {
+
+	manager.mutex.Lock()
+	defer manager.mutex.Unlock()
 
 	matchingNetworks, err := manager.dockerManager.GetNetworksByName(setupCtx, enclaveId)
 	if err != nil {
@@ -182,7 +195,52 @@ func (manager *EnclaveManager) CreateEnclave(
 	return networkId, networkIpAndMask, apiContainerId, &apiContainerIpAddr, apiContainerHostPortBinding, nil
 }
 
+func (manager *EnclaveManager) StopEnclave(ctx context.Context, enclaveId string) error {
+	manager.mutex.Lock()
+	defer manager.mutex.Unlock()
+
+	enclaveContainerSearchLabels := map[string]string{
+		enclave_object_labels.EnclaveIDContainerLabel: enclaveId,
+	}
+	allEnclaveContainers, err := manager.dockerManager.GetContainersByLabels(ctx, enclaveContainerSearchLabels, shouldKillAlreadyStoppedContainersWhenStoppingEnclave)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting containers for enclave '%v'", enclaveId)
+	}
+
+	// TODO Parallelize for perf
+	containerKillErrorStrs := []string{}
+	for _, enclaveContainer := range allEnclaveContainers {
+		containerId := enclaveContainer.GetId()
+		containerName := enclaveContainer.GetName()
+		if err := manager.dockerManager.KillContainer(ctx, containerId); err != nil {
+			wrappedContainerKillErr := stacktrace.Propagate(
+				err,
+				"An error occurred killing container '%v' with ID '%v'",
+				containerName,
+				containerId,
+			)
+			containerKillErrorStrs = append(
+				containerKillErrorStrs,
+				wrappedContainerKillErr.Error(),
+			)
+		}
+	}
+
+	var resultErr error = nil
+	if len(containerKillErrorStrs) > 0 {
+		errorStr := strings.Join(containerKillErrorStrs, "\n\n")
+		resultErr = stacktrace.NewError(
+			"One or more errors occurred killing the containers in enclave '%v':\n%v",
+			errorStr,
+		)
+	}
+	return resultErr
+
+}
+
 func (manager *EnclaveManager) DestroyEnclave(ctx context.Context, enclaveId string) error {
+	manager.mutex.Lock()
+	defer manager.mutex.Unlock()
 
 	networks, err := manager.dockerManager.GetNetworksByName(ctx, enclaveId)
 	if err != nil {